fix: move app definition to commands to fix current error in cli
